+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Until May 2022 (inclusive) no changelog was kept. We might try to reconstruct it
 * Recipe: Update uuid from ALICE recipe
 * lhapdf: update recipe and version from snddist
 * openssl: update from ALICE for new openssl versions in SLC9
+* XRootD: update recipe from ALICE, don't overwrite version in defaults
 
 ### Removed